requirements.txt
fastmcp>=2.12.4
python-dotenv
httpx
uvicorn[standard]
gunicorn